+++ lib/request.rs
@@ -1,6 +1,8 @@
 use std::fmt;
 
 use super::common::cmp_vec;
+#[cfg(feature = "cookies")]
+use super::cookie::Cookie;
 use super::header::Header;
 use super::method::Method;
 use super::query::Query;
@@ -19,6 +21,10 @@ pub struct Request {
     /// Request headers
     pub headers: Vec<Header>,
 
+    /// Request Cookies
+    #[cfg(feature = "cookies")]
+    pub cookies: Vec<Cookie>,
+
     /// Request body
     pub body: String,
 
@@ -30,6 +36,7 @@ pub struct Request {
 }
 
 impl Request {
+    #[allow(clippy::too_many_arguments)]
     /// Quick and easy way to create a request.
     ///
     /// ```rust
@@ -40,18 +47,24 @@ impl Request {
     ///    path: "/".to_string(),
     ///    query: Query::new_empty(),
     ///    headers: vec![],
+    ///    # #[cfg(feature = "cookies")]
+    ///    cookies: vec![],
     ///    body: "".to_string(),
     ///    address: "127.0.0.1:8080".to_string(),
     ///    raw_data: "".to_string(),
     /// };
     ///
-    /// assert!(request.compare(&Request::new(Method::GET, "/", Query::new_empty(), vec![], "".to_string(), "127.0.0.1:8080".to_string(), "".to_string())));
+    /// # #[cfg(feature = "cookies")]
+    /// assert!(request.compare(&Request::new(Method::GET, "/", Query::new_empty(), vec![], vec![], "".to_string(), "127.0.0.1:8080".to_string(), "".to_string())));
+    /// # #[cfg(not(feature = "cookies"))]
+    /// # assert!(request.compare(&Request::new(Method::GET, "/", Query::new_empty(), vec![], "".to_string(), "127.0.0.1:8080".to_string(), "".to_string())));
     /// ```
     pub fn new(
         method: Method,
         path: &str,
         query: Query,
         headers: Vec<Header>,
+        #[cfg(feature = "cookies")] cookies: Vec<Cookie>,
         body: String,
         address: String,
         raw_data: String,
@@ -61,6 +74,8 @@ impl Request {
             path: path.to_string(),
             query,
             headers,
+            #[cfg(feature = "cookies")]
+            cookies,
             body,
             address,
             raw_data,
@@ -89,6 +104,8 @@ impl Clone for Request {
             path: self.path.clone(),
             query: self.query.clone(),
             headers: self.headers.clone(),
+            #[cfg(feature = "cookies")]
+            cookies: self.cookies.clone(),
             body: self.body.clone(),
             address: self.address.clone(),
             raw_data: self.raw_data.clone(),
@@ -104,13 +121,15 @@ impl fmt::Debug for Request {
             headers.push_str(&format!("{}, ", header.to_string()));
         }
 
-        f.debug_struct("Request")
-            .field("method", &self.method)
-            .field("path", &self.path)
-            .field("query", &self.query)
-            .field("address", &self.address)
-            .field("headers", &headers)
-            .field("body", &self.body)
-            .finish()
+        let mut dbg = f.debug_struct("Request");
+        dbg.field("method", &self.method);
+        dbg.field("path", &self.path);
+        dbg.field("query", &self.query);
+        dbg.field("address", &self.address);
+        dbg.field("headers", &headers);
+        #[cfg(feature = "cookies")]
+        dbg.field("cookies", &self.cookies);
+        dbg.field("body", &self.body);
+        dbg.finish()
     }
 }