+++ lib/response.rs
@@ -1,3 +1,5 @@
+#[cfg(feature = "cookies")]
+use super::cookie::SetCookie;
 use super::header::Header;
 
 /// Http Response
@@ -14,7 +16,7 @@ pub struct Response {
 
 impl Response {
     /// Quick and easy way to create a response.
-    /// ## Examples
+    /// ## Example
     /// ```rust
     /// // Import Library
     /// use afire::{Response, Header};
@@ -30,7 +32,7 @@ impl Response {
     }
 
     /// Create a response from a byte Vec.
-    /// ## Examples
+    /// ## Example
     /// ```rust
     /// // Import Library
     /// use afire::{Response, Header};
@@ -49,7 +51,7 @@ impl Response {
     /// Easy way to create a successful response.
     ///
     /// Will just pass status code 200.
-    /// ## Examples
+    /// ## Example
     /// ```rust
     /// // Import Library
     /// use afire::Response;
@@ -61,6 +63,43 @@ impl Response {
     /// ```
     pub fn ok(data: &str, headers: Option<Vec<Header>>) -> Response {
         Response::new(200, data, headers.unwrap_or_default())
+    }
+
+    /// Add a cookie to a response.
+    /// ## Example
+    /// ```
+    /// // Import Library
+    /// use afire::{Response, SetCookie};
+    ///
+    /// // Create Response and add cookie
+    /// let response = Response::new(200, "🍦", vec![])
+    ///     .add_cookie(SetCookie::new("name", "value"));
+    /// ```
+    #[cfg(feature = "cookies")]
+    pub fn add_cookie(&self, cookie: SetCookie) -> Response {
+        let mut new = self.clone();
+        new.headers
+            .push(Header::new("Set-Cookie", &cookie.to_string()));
+        new.clone()
+    }
+
+    /// Add a vec of cookies to a response.
+    /// ## Example
+    /// ```
+    /// // Import Library
+    /// use afire::{Response, SetCookie};
+    ///
+    /// // Create Response and add cookie
+    /// let response = Response::new(200, "🍦", vec![])
+    ///     .add_cookies(vec![SetCookie::new("name", "value")]);
+    /// ```
+    #[cfg(feature = "cookies")]
+    pub fn add_cookies(&self, cookie: Vec<SetCookie>) -> Response {
+        for c in cookie {
+            self.add_cookie(c);
+        }
+
+        self.clone()
     }
 }