+++ Changelog.md
@@ -7,6 +7,10 @@
 - Auto decode url encoded Querys
 - Update Readme
 - Proper Spelling In Description
+- Add support for Request Cookies
+- Add Cookies to a Feature
+- Add Support for Response Cookies
+- Deprecate .any routes
 
 
 # 0.1.4